Refactors config as app and services
