--- config/lib/middleware/authenticate.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const configVars = {
-  auth: {
-    name: process.env.DS_GAMBIT_ADMIN_BASIC_AUTH_NAME || 'puppet',
-    pass: process.env.DS_GAMBIT_ADMIN_BASIC_AUTH_PASS || 'sloth',
-  },
-  unauthorizedErrorMessage: 'Invalid or missing auth name and pass.',
-};
-
-module.exports = configVars;
