--- config/lib/northstar.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  apiKey: process.env.DS_NORTHSTAR_API_KEY,
-  baseUri: process.env.DS_NORTHSTAR_API_BASEURI,
-};
