+++ lib/gambit-campaigns.js
@@ -2,7 +2,7 @@
 
 const superagent = require('superagent');
 const logger = require('heroku-logger');
-const config = require('../config/lib/gambit-campaigns');
+const config = require('../config/services').gambitContent;
 
 /**
  * @param {string} path
