+++ lib/gambit-conversations.js
@@ -2,7 +2,7 @@
 
 const superagent = require('superagent');
 const logger = require('heroku-logger');
-const config = require('../config/lib/gambit-conversations');
+const config = require('../config/services').gambitConversations;
 
 /**
  * @param {object} apiRes - Gambit Conversations v1 api response
